backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
pytesseract==0.3.10
deep-translator==1.11.4
requests==2.31.0
pydantic==2.7.1
aiofiles==23.2.1
